+++ src/urldecode.h
@@ -1,9 +1,12 @@
 #pragma once
 
 #include <Rcpp.h>
+
+#include <string>
+
 using namespace Rcpp;
 
-std::string decode(String u);
-CharacterVector Rcpp_url_decode2(CharacterVector& url);
+std::string decode(const char* input, size_t len);
+CharacterVector Rcpp_url_decode2(const CharacterVector& url);
 std::string str_reverse(std::string x);
-CharacterVector url_reverse(CharacterVector& urls);
+CharacterVector url_reverse(const CharacterVector& urls);